[improvement](hashjoin) fix problem that fragment instances are not necessarily synchronized for broadcast hash join
